requirements.txt
# netfilterqueue>=0.8.1  # Linux only - will use mock version on other systems
dpkt>=1.9.8
pyahocorasick>=2.0.0
flask>=3.1.1
pyyaml>=6.0
